Synchronously update the workload status when the pod changes.
